update api to update colors

+++ src/app/api/products/[productId]/route.ts
@@ -12,16 +12,24 @@ export async function DELETE(
       return new NextResponse("Unauthorized", { status: 401 })
     }
 
-    const product = await prisma.product.delete({
-      where: {
-        id: params.productId,
-      },
+    // Ensure the product belongs to the user
+    const product = await prisma.product.findUnique({
+      where: { id: params.productId },
     })
 
-    return NextResponse.json(product)
+    if (!product) {
+      return new NextResponse("Product not found", { status: 404 })
+    }
+
+    // Delete the product
+    await prisma.product.delete({
+      where: { id: params.productId },
+    })
+
+    return new NextResponse("Product deleted successfully", { status: 200 })
   } catch (error) {
-    console.error("Error when deletting product", error)
-    return new NextResponse("Internal server", { status: 500 })
+    console.error("Error deleting product", error)
+    return new NextResponse("Internal server error", { status: 500 })
   }
 }
 
@@ -30,14 +38,22 @@ export async function PATCH(
   { params }: { params: { productId: string } }
 ) {
   try {
-    if (!params.productId) {
-      return new NextResponse("Product Id is required", { status: 400 })
-    }
-
     const userId = req.headers.get("X-USER-ID")
 
     if (!userId) {
       return new NextResponse("Unauthorized", { status: 401 })
+    }
+
+    if (!params.productId) {
+      return new NextResponse("Product Id is required", { status: 400 })
+    }
+
+    const product = await prisma.product.findUnique({
+      where: { id: params.productId },
+    })
+
+    if (!product) {
+      return new NextResponse("Product not found", { status: 404 })
     }
 
     const {
@@ -47,51 +63,60 @@ export async function PATCH(
       discount,
       stock,
       categoryId,
-      colorId,
+      colors,
       isFeatured,
       description,
       isAvailable,
     } = await req.json()
 
-    const categories = await prisma.category.findMany({
+    // Handle colors update
+
+    const colorId = await prisma.color.findMany({
       where: {
-        id: categoryId,
+        value: {
+          in: colors,
+        },
       },
     })
 
-    for (const category of categories) {
-      await prisma.category.update({
-        where: {
-          id: category.id,
-        },
-        data: { title: category.title },
-      })
-    }
+    console.log(colorId)
 
-    const product = await prisma.product.update({
-      where: {
-        id: params.productId,
-      },
+    const imageUrls = images.map((image: string | { url: string }) =>
+      typeof image === "string" ? image : image.url
+    )
+
+    const updatedProduct = await prisma.product.update({
+      where: { id: params.productId },
       data: {
         title,
         price,
         discount,
         stock,
         isFeatured,
-        colorId,
         isAvailable,
         description,
-        categories: {
-          connect: categories.map((cat) => ({ id: cat.id })),
+        colors: {
+          set: [],
+          connect: colorId.map((color) => ({ id: color.id })),
         },
-        images: [...images.map((image: { url: string }) => image)],
+        categories: {
+          set: [], // First disconnect all existing categories
+          connect: categoryId ? [{ id: categoryId }] : [],
+        },
+        images: imageUrls,
+      },
+      include: {
+        colors: true, // Include colors in the response
+        categories: true, // Include categories in the response
       },
     })
 
-    return NextResponse.json(product)
+    console.log(updatedProduct)
+
+    return NextResponse.json(updatedProduct)
   } catch (error) {
-    console.error("[PRODUCT_PATCH]", error)
-    return new NextResponse("Internal error", { status: 500 })
+    console.error("Error updating product", error)
+    return new NextResponse("Internal server error", { status: 500 })
   }
 }
 
@@ -107,18 +132,20 @@ export async function GET(
     }
 
     if (!params.productId) {
-      return new NextResponse("Product id is required", { status: 400 })
+      return new NextResponse("Product Id is required", { status: 400 })
     }
 
-    const product = await prisma.product.findUniqueOrThrow({
-      where: {
-        id: params.productId,
-      },
+    const product = await prisma.product.findUnique({
+      where: { id: params.productId },
     })
+
+    if (!product) {
+      return new NextResponse("Product not found", { status: 404 })
+    }
 
     return NextResponse.json(product)
   } catch (error) {
-    console.error("Error when getting product", error)
-    return new NextResponse("Internal error", { status: 500 })
+    console.error("Error retrieving product", error)
+    return new NextResponse("Internal server error", { status: 500 })
   }
 }
